requirements.txt
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Dance==7.1.0
Flask-PyMongo==3.0.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
protobuf==6.31.0
proto-plus==1.26.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pymongo==4.13.0
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
tomlkit==0.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
gunicorn==22.0.0
